Réorganisation + Check warnings
