+++ compose/popup.css
@@ -18,8 +18,8 @@ h2 {
     text-align: center;
     font-weight: bold;
     font-size: 11pt;
-    margin-bottom: 0px;
-    margin-top: 0px;
+    margin-bottom: 0;
+    margin-top: 0;
 }
 
 
@@ -62,7 +62,7 @@ td {
 #emailSizeTable,
 #equivalencesTable {
     width: 100%;
-    margin-top: 0px;
+    margin-top: 0;
 }
