+++ compose/popup.html
@@ -4,21 +4,21 @@
 <head>
     <meta charset="utf-8">
     <title>Estimation de l'impact environnemental de l'envoi de ce courriel</title>
-    <link rel="stylesheet" type="text/css" media="screen" href="style/popup.css">
-    <link rel="stylesheet" type="text/css" media="screen" href="style/tooltip.css">
-    <script src="scripts/popup.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="popup.css">
+    <link rel="stylesheet" type="text/css" media="screen" href="../global/tooltip.css">
+    <script src="popup.js" type="module"></script>
 </head>
 
 <body>
     <div class="popup-title">
         <div id="preferencesButton">
-            <a href="preferences/preferences.html" id="preferencesLink">
-                <img src="images/settings-icon.png" alt="Préférences" title="Préférences" height="15px" />
+            <a href="../preferences/preferences.html" id="preferencesLink">
+                <img src="../images/settings-icon.png" alt="Préférences" title="Préférences" height="15px" />
             </a>
         </div>
         <h1>Estimation de l'impact environnemental<br> global de ce courriel</h1>
         <div class='tooltip tooltip-left' id="infoButton">
-            <img src="images/info-icon.png" alt="Plus d'informations..." title="Plus d'informations..." height="15px" />
+            <img src="../images/info-icon.png" alt="Plus d'informations..." title="Plus d'informations..." height="15px" />
             <span class='tooltiptext tooltiptext-left' style="width: 290px;">
             <small>
                 Sont pris en compte : la fabrication des équipements électriques et électroniques (EEE), l'envoi et le stockage, la transmission, le traitement, la réception, l'utilisation, et la fin de vie des EEE.<br><br>
@@ -39,26 +39,26 @@
         <div id="sec-email">
             <table id="emailSizeTable">
                 <tr>
-                    <td><img src="images/user-icon_gray.png" alt="User icon" title="User" height="16px" /></td>
+                    <td><img src="../images/user-icon_gray.png" alt="User icon" title="User" height="16px" /></td>
                     <td class="gray"> Taille de l'en-tête (approximative)
                         <div id="recipientsWarning" class="tooltip tooltip-left" style="margin-left: 4px;"></div>
                     </td>
                     <td class="gray" id="header-size">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/text-icon_gray.png" alt="Text icon" title="Text" height="16px" /></td>
+                    <td><img src="../images/text-icon_gray.png" alt="Text icon" title="Text" height="16px" /></td>
                     <td class="gray">Taille du corps de texte</td>
                     <td class="gray" id="body-size">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/attachment-icon_gray.png" alt="Attachment icon" title="Attachment" height="16px" /></td>
+                    <td><img src="../images/attachment-icon_gray.png" alt="Attachment icon" title="Attachment" height="16px" /></td>
                     <td class="gray">Taille des pièces jointes
                         <div id="attachmentWarning" class="tooltip tooltip-left" style="margin-left: 4px;"></div>
                     </td>
                     <td class="gray" id="attachments-size">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/email-icon.png" alt="Email icon" title="Email" height="16px" /></td>
+                    <td><img src="../images/email-icon.png" alt="Email icon" title="Email" height="16px" /></td>
                     <td>Taille totale</td>
                     <td id="size">Calcul...</td>
                 </tr>
@@ -71,32 +71,32 @@
             <h2>Équivalences</h2>
             <table id="equivalencesTable">
                 <tr>
-                    <td><img src="images/co2-icon.png" alt="CO₂ icon" title="CO₂" height="16px" /></td>
+                    <td><img src="../images/co2-icon.png" alt="CO₂ icon" title="CO₂" height="16px" /></td>
                     <td>Émission de CO<sub>2</sub>e</td>
                     <td id="co2">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/petroleum-icon.png" alt="Petroleum icon" title="Petroleum" height="16px" /></td>
+                    <td><img src="../images/petroleum-icon.png" alt="Petroleum icon" title="Petroleum" height="16px" /></td>
                     <td>Consommation de pétrole</td>
                     <td id="oil">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/car-icon.png" alt="Car icon" title="Car" height="16px" /></td>
+                    <td><img src="../images/car-icon.png" alt="Car icon" title="Car" height="16px" /></td>
                     <td>Distance parcourue en voiture</td>
                     <td id="car">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/tgv-icon.png" alt="TGV icon" title="TGV" height="16px" /></td>
+                    <td><img src="../images/tgv-icon.png" alt="TGV icon" title="TGV" height="16px" /></td>
                     <td>Distance parcourue en TGV</td>
                     <td id="tgv">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/bulb-icon.png" alt="Lightbulb icon" title="Lightbulb" height="16px" /></td>
+                    <td><img src="../images/bulb-icon.png" alt="Lightbulb icon" title="Lightbulb" height="16px" /></td>
                     <td id="bulbw">Temps d'utilisation d'une ampoule de </td>
                     <td id="bulb">Calcul...</td>
                 </tr>
                 <tr>
-                    <td><img src="images/lungs-icon.png" alt="Lungs icon" title="Lungs" height="16px" /></td>
+                    <td><img src="../images/lungs-icon.png" alt="Lungs icon" title="Lungs" height="16px" /></td>
                     <td>Durée de respiration</td>
                     <td id="breathing">Calcul...</td>
                 </tr>
